// PolicyManagementAngular/src/app/app-entites/models/inspection/inspection-search-policy-model.ts
export interface IInspectionSearchPolicyModel {
    BranchId: number;
    ControlNumber: string;
    CustomerName: string;
    CustomerId?: number;
    InsuranceCompanyId?: number;
    PosId?: number;
    PolicyNumber: string;
    PolicyStartFromDate: string;
    PolicyStartToDate: string;
    RegistrationNumber: string;
    ManufacturerId?: number;
    ModelId?: number;
    CustomerPhone: string;
}